+++ apps/www/app/routes.ts
@@ -56,6 +56,20 @@ export default [
       route('/components', 'routes/components/components.tsx', {
         id: 'components',
       }),
+      layout('./layouts/components/layout.tsx', [
+        ...prefix('/changelog', [
+          route('/:package', 'routes/components/changelog.tsx', {
+            id: 'changelog-page',
+          }),
+        ]),
+        ...prefix('/components', [
+          ...prefix(':component', [
+            route('/*', 'routes/components/component.tsx', {
+              id: 'components-page',
+            }),
+          ]),
+        ]),
+      ]),
       route('*', 'routes/not-found.tsx', {
         id: 'not-found-lang',
       }),